requirements.txt
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.2
asttokens==2.0.5
attrs==21.4.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==4.1.0
certifi==2022.6.15
cffi==1.15.0
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
Django==4.1
entrypoints==0.4
executing==0.8.3
fastjsonschema==2.15.1
imbalanced-learn==0.9.0
ipykernel==6.9.1
ipython==8.4.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.2.2
jupyter-core==4.10.0
jupyterlab-pygments==0.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.2
mistune==0.8.4
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
notebook==6.4.11
numpy==1.23.1
packaging==21.3
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
pip==22.1.2
prometheus-client==0.13.1
prompt-toolkit==3.0.20
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.4
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2022.1
PyYAML==5.4.1
pyzmq==23.2.0
scikit-learn==1.0.2
scipy==1.8.1
Send2Trash==1.8.0
setuptools==61.2.0
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
terminado==0.13.1
testpath==0.6.0
threadpoolctl==3.1.0
tornado==6.1
traitlets==5.1.1
typing_extensions==4.1.1
wcwidth==0.2.5
webencodings==0.5.1
wheel==0.37.1
